move is_guarded from move.h to guard.h
Conflicts:
	src/move.c
