+++ src/alchemy.c
@@ -19,6 +19,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <platform.h>
 #include <kernel/config.h>
 #include "alchemy.h"
+#include "guard.h"
 #include "move.h"
 #include "skill.h"
 #include "study.h"
