+++ src/guard.c
@@ -20,7 +20,10 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <platform.h>
 #include <kernel/config.h>
 #include "guard.h"
+#include "laws.h"
+#include "monster.h"
 
+#include <kernel/ally.h>
 #include <kernel/save.h>
 #include <kernel/unit.h>
 #include <kernel/faction.h>
@@ -144,4 +147,80 @@ void guard(unit * u, unsigned int mask)
 {
     unsigned int flags = guard_flags(u);
     setguard(u, flags & mask);
+}
+
+static bool is_guardian_u(const unit * guard, unit * u, unsigned int mask)
+{
+    if (guard->faction == u->faction)
+        return false;
+    if (is_guard(guard, mask) == 0)
+        return false;
+    if (alliedunit(guard, u->faction, HELP_GUARD))
+        return false;
+    if (ucontact(guard, u))
+        return false;
+    if (!cansee(guard->faction, u->region, u, 0))
+        return false;
+    if (!(u_race(guard)->flags & RCF_FLY) && u_race(u)->flags & RCF_FLY)
+        return false;
+
+    return true;
+}
+
+static bool is_guardian_r(const unit * guard)
+{
+    if (guard->number == 0)
+        return false;
+    if (besieged(guard))
+        return false;
+
+    /* if region_owners exist then they may be guardians: */
+    if (guard->building && rule_region_owners() && guard == building_owner(guard->building)) {
+        faction *owner = region_get_owner(guard->region);
+        if (owner == guard->faction) {
+            building *bowner = largestbuilding(guard->region, &cmp_taxes, false);
+            if (bowner == guard->building) {
+                return true;
+            }
+        }
+    }
+
+    if ((guard->flags & UFL_GUARD) == 0)
+        return false;
+    return fval(u_race(guard), RCF_UNARMEDGUARD) || is_monsters(guard->faction) || (armedmen(guard, true) > 0);
+}
+
+bool is_guard(const struct unit * u, unsigned int mask)
+{
+    return is_guardian_r(u) && (getguard(u) & mask) != 0;
+}
+
+unit *is_guarded(region * r, unit * u, unsigned int mask)
+{
+    unit *u2;
+    int noguards = 1;
+
+    if (!fval(r, RF_GUARDED)) {
+        return NULL;
+    }
+
+    /* at this point, u2 is the last unit we tested to
+    * be a guard (and failed), or NULL
+    * i is the position of the first free slot in the cache */
+
+    for (u2 = r->units; u2; u2 = u2->next) {
+        if (is_guardian_r(u2)) {
+            noguards = 0;
+            if (is_guardian_u(u2, u, mask)) {
+                /* u2 is our guard. stop processing (we might have to go further next time) */
+                return u2;
+            }
+        }
+    }
+
+    if (noguards) {
+        /* you are mistaken, sir. there are no guards in these lands */
+        freset(r, RF_GUARDED);
+    }
+    return NULL;
 }
