+++ src/move.h
@@ -60,8 +60,6 @@ extern "C" {
     int personcapacity(const struct unit *u);
     void movement(void);
     void run_to(struct unit *u, struct region *to);
-    struct unit *is_guarded(struct region *r, struct unit *u, unsigned int mask);
-    bool is_guard(const struct unit *u, unsigned int mask);
     int enoughsailors(const struct ship *sh, int sumskill);
     bool canswim(struct unit *u);
     bool canfly(struct unit *u);
